requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
django-uni-form==0.9.0
Pillow==10.0.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
